refactor: config file
