+++ src/main/java/com/example/swip/config/security/JwtIssuer.java
@@ -1,7 +1,6 @@
-package com.example.swip.config;
+package com.example.swip.config.security;
 
 import com.auth0.jwt.JWT;
-import com.auth0.jwt.JWTCreator;
 import com.auth0.jwt.algorithms.Algorithm;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
