+++ src/main/java/com/example/swip/config/security/UserPrincipalAuthenticationToken.java
@@ -1,4 +1,4 @@
-package com.example.swip.config;
+package com.example.swip.config.security;
 
 import org.springframework.security.authentication.AbstractAuthenticationToken;
