+++ src/main/java/com/example/swip/service/impl/ChatServerServiceImpl.java
@@ -1,5 +1,6 @@
 package com.example.swip.service.impl;
 
+import com.example.swip.config.HttpRequestSender;
 import com.example.swip.dto.DefaultResponse;
 import com.example.swip.dto.chat.ChatProfileRequest;
 import com.example.swip.dto.study.PostStudyAddMemberRequest;
@@ -13,16 +14,11 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
-import java.io.*;
-import java.net.HttpURLConnection;
-import java.net.URL;
-import java.net.URLEncoder;
-import java.nio.charset.StandardCharsets;
-import java.util.Map;
-
 @Service
 @RequiredArgsConstructor
 public class ChatServerServiceImpl implements ChatServerService {
+    private final HttpRequestSender httpRequestSender;
+
     @Value("${swyp.chat.server.uri}")
     private String reqUserURL;
     public Pair<String, Integer> postUser(ChatProfileRequest chatProfileRequest){
@@ -30,7 +26,7 @@ public class ChatServerServiceImpl implements ChatServerService {
         ObjectMapper objectMapper = new ObjectMapper();
         try {
             String jsonInputString = objectMapper.writeValueAsString(chatProfileRequest);
-            result = sendHttpRequest(reqUserURL, "POST", jsonInputString, null);
+            result = httpRequestSender.sendHttpRequest(reqUserURL, "POST", jsonInputString, null);
         } catch (JsonProcessingException e) {
             e.printStackTrace();
             return Pair.of("Failed to convert object to JSON", 500);
@@ -46,7 +42,7 @@ public class ChatServerServiceImpl implements ChatServerService {
             if (!reqUserURL.isEmpty())
                 reqUpdateUserURL = String.format("%s/%s", reqUserURL, chatProfileRequest.getPk());
             String jsonInputString = objectMapper.writeValueAsString(chatProfileRequest);
-            result = sendHttpRequest(reqUpdateUserURL, "PATCH", jsonInputString, null);
+            result = httpRequestSender.sendHttpRequest(reqUpdateUserURL, "PATCH", jsonInputString, null);
         }catch (JsonProcessingException e) {
             e.printStackTrace();
             return Pair.of("Failed to convert object to JSON", 500);
@@ -57,7 +53,7 @@ public class ChatServerServiceImpl implements ChatServerService {
         String reqDeleteUserURL = "";
         if(!reqUserURL.isEmpty())
             reqDeleteUserURL = String.format("%s/%s", reqUserURL, userId);
-        Pair<String, Integer> result = sendHttpRequest(reqDeleteUserURL, "DELETE", null, null);
+        Pair<String, Integer> result = httpRequestSender.sendHttpRequest(reqDeleteUserURL, "DELETE", null, null);
         return result;
     }
 
@@ -71,7 +67,7 @@ public class ChatServerServiceImpl implements ChatServerService {
         String jsonInputString = "{\"studyId\":\""+ postStudyRequest.getStudyId().toString()
                 +"\",\"pk\":\""+postStudyRequest.getPk().toString()
                 +"\",\"name\":\""+postStudyRequest.getName()+"\"}";
-        Pair<String, Integer> result = sendHttpRequest(studyReqURL, "POST", jsonInputString, null);
+        Pair<String, Integer> result = httpRequestSender.sendHttpRequest(studyReqURL, "POST", jsonInputString, null);
 
         return DefaultResponse.builder()
                 .message(result.getFirst())
@@ -88,7 +84,7 @@ public class ChatServerServiceImpl implements ChatServerService {
                 +"\",\"userId\":\""+postStudyAddmemberRequest.getUserId().toString()
                 +"\",\"type\":\""+postStudyAddmemberRequest.getType()+"\"}";
 
-        Pair<String, Integer> result = sendHttpRequest(studyAddMemberReqURL, "PUT", jsonInputString, bearerToken);
+        Pair<String, Integer> result = httpRequestSender.sendHttpRequest(studyAddMemberReqURL, "PUT", jsonInputString, bearerToken);
         return DefaultResponse.builder()
                 .message(result.getFirst())
                 .build();
@@ -102,7 +98,7 @@ public class ChatServerServiceImpl implements ChatServerService {
 
         String jsonInputString = "{\"studyId\":\""+ postStudymemberRequest.getStudyId().toString()
                 +"\",\"userId\":\""+postStudymemberRequest.getUserId().toString() +"\"}";
-        Pair<String, Integer> result = sendHttpRequest(studyDeleteMemberReqURL, "PUT", jsonInputString, bearerToken);
+        Pair<String, Integer> result = httpRequestSender.sendHttpRequest(studyDeleteMemberReqURL, "PUT", jsonInputString, bearerToken);
 
         System.out.println("jsonInputString = " + jsonInputString);
         System.out.println("bearerToken = " + bearerToken);
@@ -110,70 +106,5 @@ public class ChatServerServiceImpl implements ChatServerService {
         return DefaultResponse.builder()
                 .message(result.getFirst())
                 .build();
-    }
-
-    public static Pair<String, Integer> sendHttpRequest(String reqURL, String method, String jsonInputString, String bearerToken) {
-        StringBuilder response = new StringBuilder();
-        if(reqURL==null || reqURL.isEmpty())
-            return Pair.of("deprecated",200);
-        try {
-            URL url = new URL(reqURL);
-            HttpURLConnection conn = (HttpURLConnection) url.openConnection();
-
-            // HTTP 메서드 설정
-            conn.setRequestMethod(method);
-            conn.setRequestProperty("Content-Type", "application/json");
-            if(bearerToken != null) {
-                conn.setRequestProperty("Authorization", "Bearer " + bearerToken); //Bearer Token 추가
-            }
-            conn.setDoOutput(true);
-            conn.setConnectTimeout(3000); // 1초
-            conn.setReadTimeout(3000);
-
-            if ("POST".equals(method) || "PUT".equals(method)) {
-                try (OutputStream os = conn.getOutputStream()) {
-                    byte[] input = jsonInputString.getBytes("utf-8"); // JSON 문자열을 바이트 배열로 변환
-                    os.write(input, 0, input.length); // 변환된 바이트 배열을 출력 스트림을 통해 전송
-                }
-            }
-
-            int responseCode = conn.getResponseCode();
-            System.out.println("responseCode : " + responseCode);
-
-            if (responseCode >= 300 || responseCode < 200) {
-                return Pair.of("HTTP request failed. Response code: ", responseCode);
-            }
-
-            // 요청을 통해 얻은 Response 메세지 읽어오기
-            try (BufferedReader br = new BufferedReader(new InputStreamReader(conn.getInputStream()))) {
-                String line;
-                while ((line = br.readLine()) != null) {
-                    response.append(line);
-                }
-            }
-            System.out.println("response body : " + response);
-            return Pair.of("success!", 200);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
-        return Pair.of("time out", 408);
-    }
-    public static Pair<String, Integer> sendRequestUserQueryParam(String reqURL, String method, Map<String, String> queryParams, String bearerToken) {
-        String jsonInputString = null; // GET 요청에는 body가 없으므로 null 처리
-        // 쿼리 파라미터 추가
-        if (queryParams != null && !queryParams.isEmpty()) {
-            StringBuilder query = new StringBuilder();
-            for (Map.Entry<String, String> entry : queryParams.entrySet()) {
-                if (query.length() > 0) {
-                    query.append("&");
-                }
-                query.append(URLEncoder.encode(entry.getKey(), StandardCharsets.UTF_8));
-                query.append("=");
-                query.append(URLEncoder.encode(entry.getValue(), StandardCharsets.UTF_8));
-            }
-            reqURL += "?" + query.toString();
-        }
-        return sendHttpRequest(reqURL, method, jsonInputString, bearerToken);
     }
 }
